# sdk/trace: optimize recordingSpan endTime with atomics